requirements.txt
pandas>=2.0
openpyxl>=3.1
requests>=2.28
selenium>=4.10
schedule>=1.2
python-dotenv>=1.0
